Part 6
Refactor any data update into ServerActualFireSystem.cs. This allows me to use GasAnalyzerSystem.cs to do some jank that it was never intended for, but works great for my use case. GetOxidation should work now™, so now it's just figure out a way to light the fire, spawn it, and populate it's data.
